moving models to src
