--- src/db/models/ingredients.js
@@ -1,7 +1,0 @@
-import { Schema, model } from 'mongoose';
-
-const ingredientsSchema = new Schema({
-  name: { type: String, required: true, unique: true },
-});
-
-export const IngredientsCollection = model('ingredients', ingredientsSchema);
