--- src/db/models/user.js
@@ -1,26 +1,0 @@
-import { model, Schema } from 'mongoose';
-
-const usersSchema = new Schema(
-  {
-    name: { type: String, required: true },
-    email: { type: String, required: true, unique: true },
-    password: { type: String, required: true },
-    favouriteRecipes: [
-      {
-        type: Schema.Types.ObjectId,
-        ref: 'recipes',
-      },
-    ],
-  },
-  { timestamps: true, versionKey: false },
-);
-
-usersSchema.index({ favouriteRecipes: 1 });
-
-usersSchema.methods.toJSON = function () {
-  const obj = this.toObject();
-  delete obj.password;
-  return obj;
-};
-
-export const UsersCollection = model('users', usersSchema);
