+++ src/models/category.js
@@ -1,0 +1,7 @@
+import { Schema, model } from 'mongoose';
+
+const categoriesSchema = new Schema({
+  name: { type: String, required: true, unique: true },
+});
+
+export const CategoriesCollection = model('categories', categoriesSchema);
