+++ src/models/ingredient.js
@@ -1,0 +1,7 @@
+import { Schema, model } from 'mongoose';
+
+const ingredientsSchema = new Schema({
+  name: { type: String, required: true, unique: true },
+});
+
+export const IngredientsCollection = model('ingredients', ingredientsSchema);
